using Angular CSRF protection
By default, AngularJS provides a mechanism to combat cross-site
scripting attacks. Any request that uses the $http service will read a
token set in a cookie by the key of XSRF-TOKEN and set it as a header
X-XSRF-TOKEN. Since browsers only enable a site by the same domain the
ability to read the cookie, it will ensure that the request comes from
HTML running on the same domain.
To enable this protection in our AngularJS app, we’ll need to add some
middleware in our Rails app. We’ll add an after_filter to our
ApplicationController